requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
cffi==2.0.0
cross-web==0.4.1
dj-database-url==3.1.0
Django==6.0.2
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-q2==1.7.6
django-extensions==3.2.3
djangorestframework==3.14.0
et_xmlfile==2.0.0
graphql-core==3.2.7
openpyxl==3.1.2
packaging==26.0
Pillow==10.1.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
qrcode==7.4.2
pytz==2025.2
redis==5.0.1
reportlab==4.2.5
six==1.17.0
sqlparse==0.5.5
strawberry-graphql==0.288.0
strawberry-graphql-django==0.74.2
setuptools<81
typing_extensions==4.15.0
tzdata==2025.3